Tell the libraries API how to detect the basex-api.
Previously, the basex-api was not being detected at all, however the
status report entry that would have indicated this was being stomped by
the `islandora_cwrc_writer` module, because both modules used the same
requirements key.

Note that this module requires that we download only a single file from
the main library and put it in a different location than it would be if
we were to download the full library. This is not ideal for a number of
reasons:

* we cannot accurately determine which version of the file we have so we
  cannot tell if it requires security updates, or if the version we are
  downloading will work or not,
* if the file is changed upstream in the future to require other files
  in the same directory (`require` / `include` / PHP autoloading), then
  this will break.
* it requires extra work to set up / update,
* setting up this library is totally different from setting up other
  libraries (breaks the conventions of Drupal / the Libraries API),

... however, I don't know all the motivations for doing it this way, and
I believe fixing this is beyond the scope of this pull request, and the
changes needed to fix this module will make upgrading difficult, or, at
the very least, non-trivial; and will conflict with this pull request. I
think these larger changes are best left for a future pull request.